\name{pantone-package}
\alias{pantone-package}
\alias{pantone}
\docType{package}
\title{
Pantone colors of the year (individual and palette)
}
\description{
Pantone colors of the year (individual and palette)
}
\details{
\tabular{ll}{
Package: \tab pantone\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-05\cr
License: \tab MIT License\cr
}
Pantone colors of the year (individual and palette)
}
\author{
Bob Rudis (@hrbrmstr)

Maintainer: Bob Rudis <bob@rudis.net>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
